Description added
We can prove to be a valuable additional resource supporting your
company secretarial function in-house as well as representing a cost
effective corporate handler of corporate legal services, providing a
vital guidance and advice to the company and its directors on the
various legal responsibilities of the same in various geographical
locations fulfilling the responsibilities through accurate application
of governance and compliance.
A company Secretary is a vital part of the organisation and we play that
role for you in a highly competitive manner with a cost effective means
handling the compliance work and advising you on various corporate
compliances that your organisation needs.
Company secretaries keep directors and shareholders informed of their
legal responsibilities and make sure main policies relating to the
governance of the company are up to date and accurate. Hiring a
Practicing Company secretary helps you save the trouble and costs
related to employing one in case you do not fall under the category of
business that needs a company secretary in full time employment.
The benefits of a company secretary are widely felt because it is vital
to ensure organisations of all sizes can achieve their business goals in
a legitimate manner. We also offer Corporate Governance services to
ensure that your corporate governance framework is delivering your
strategic objectives.