ST form uploaded
F. No. 137/22/2012-Service Tax ( Part-I)
Government of India
Ministry of Finance
Department of Revenue
Central Board of Excise and Customs
Service Tax Wing

New Delhi, the 28th February, 2015
ORDER No. 1/2015-SERVICE TAX
The legal provisions for registration in service tax are contained in
section 69 of the
Finance Act, 1994, rule 4 of the Service Tax Rules, 1994 and the Service
Tax (Registration of
Special Category of Persons) Rules, 2005. Paragraph 2 of Circular
97/8/2007-Service Tax dated
23-8-2007 and Order No. 2/2011-Service Tax dated 13-12-2011 also explain
some of the
procedural aspects of registration in service tax.
In supercession of Order No. 2/2011-Service Tax dated 13-12-2011, the
Central Board of
Excise and Customs specifies the following documentation, time limits
and procedure with
respect to filing of registration applications for single premises,
which shall come into effect
from 1-3-2015.